+++ Editor/BlendshapeSyncAnimationProcessor.cs
@@ -193,22 +193,22 @@ namespace nadena.dev.modular_avatar.core.editor
                     else
                     {
                         const float epsilon = 0.005f; // ~200fps
-                        var remappedCurve = new AnimationCurve();
-                        foreach (var key in curve.keys)
+                        var keys = curve.keys;
+                        foreach (ref var key in keys.AsSpan())
                         {
                             var t = Mathf.Clamp01(key.value / 100f);
                             var val = remapCurve.Evaluate(t) * 100f;
                             var tPlus = Mathf.Clamp01(t + epsilon);
                             var tMinus = Mathf.Clamp01(t - epsilon);
-                            var valPlus = remapCurve.Evaluate(tPlus) * 100f;
-                            var valMinus = remapCurve.Evaluate(tMinus) * 100f;
+                            var valPlus = remapCurve.Evaluate(tPlus);
+                            var valMinus = remapCurve.Evaluate(tMinus);
                             var slope = (valPlus - valMinus) / (tPlus - tMinus);
-                            remappedCurve.AddKey(new Keyframe(
-                                key.time, val,
-                                key.inTangent * slope,
-                                key.outTangent * slope
-                            ));
+
+                            key.value = val;
+                            key.inTangent *= slope;
+                            key.outTangent *= slope;
                         }
+                        var remappedCurve = new AnimationCurve(keys);
                         clip.SetFloatCurve(dst.ToEditorCurveBinding(asc), remappedCurve);
                     }
                 }